+++ src/AlgorithmsInterface.jl
@@ -18,12 +18,14 @@ include("interface/state.jl")
 include("interface/interface.jl")
 
 include("stopping_criterion.jl")
+include("default_state.jl")
 include("logging.jl")
 
 include("test_suite.jl")
 
 # general interface
 export Algorithm, Problem, State
+export DefaultState
 export initialize_state, initialize_state!
 export finalize_state!